Refactoring: In order to incorporate custom processes (like in git hub challenge) into agent's behavior, a new method customAgentLoop() was added to DASHAction class. It is None be default, but it can be defined in subclasses.
In current implementation, DASHAction.agentLoop() first tries to call customAgentLoop() if it id defined. If customAgentLoop() is not defined it uses S1 and S2 processes as usual. Also if customAgentLoop() returns True, DASHAction.agentLoop() uses S1 and S2 processes in the same loop/step. In that way both custom agent process and S1 and S2 can augment each other or replace each other if necessary.

To create a new agent with a custom process, see example in git_user_agent.